Add files via upload
# Here are the 4 versions, with their own characteristics:
## First Version: 
Performs addition, subtraction, multiplication, and division by checking first asking the 2 numbers used, and finally uses checks the operator.
After a single calculation, the program stops.

## Second Version:
Capable of using the previous result and perform the same operations to form a new response. 
The program will be in a loop until an error is encountered of "0" is inputted as the operator.
Now, the program checks if the operator is one of the valid 4 operators, and if not, then will terminate.

## Third Version:
This iteration will now repeat the whole program if an incorrect operator is given. This allows the calculator
to make both new calculations and build upon previous answers. Furthermore, the code is greatly simplified
from previous iterations.

## Fourth Version (Latest):
This time, the only change was making the program handle the error created by giving the program any symbol
that is not a number whenever it asks for a number for the first time.